references/requirements-py312.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.8.0
faiss-cpu>=1.8.0.post1
tiktoken>=0.7.0
